+++ aws-xray-propagator/src/test/java/io/opentelemetry/contrib/awsxray/propagator/internal/AwsComponentProviderTest.java
@@ -25,7 +25,7 @@ class AwsComponentProviderTest {
             + "propagator:\n"
             + "  composite:\n"
             + "    - xray:\n"
-            + "    - xray-lambda:\n";
+            + "    - xray_lambda:\n";
 
     OpenTelemetrySdk openTelemetrySdk =
         DeclarativeConfiguration.parseAndCreate(