+++ inferred-spans/src/main/java/io/opentelemetry/contrib/inferredspans/internal/SamplingProfiler.java
@@ -150,6 +150,7 @@ public class SamplingProfiler implements Runnable {
   private final ProfilingActivationListener activationListener;
 
   private final Supplier<Tracer> tracerProvider;
+  @Nullable private final File tempDir;
 
   private final AsyncProfiler profiler;
   @Nullable private volatile Future<?> profilingTask;
@@ -165,14 +166,17 @@ public class SamplingProfiler implements Runnable {
    * @param activationEventsFile activation events file, if {@literal null} a temp file will be used
    * @param jfrFile java flight recorder file, if {@literal null} a temp file will be used instead
    */
+  @SuppressWarnings("this-escape")
   public SamplingProfiler(
       InferredSpansConfiguration config,
       SpanAnchoredClock nanoClock,
       Supplier<Tracer> tracerProvider,
       @Nullable File activationEventsFile,
-      @Nullable File jfrFile) {
+      @Nullable File jfrFile,
+      @Nullable File tempDir) {
     this.config = config;
     this.tracerProvider = tracerProvider;
+    this.tempDir = tempDir;
     this.scheduler =
         Executors.newSingleThreadScheduledExecutor(
             r -> {
@@ -252,12 +256,13 @@ public class SamplingProfiler implements Runnable {
 
   private synchronized void createFilesIfRequired() throws IOException {
     if (jfrFile == null || !jfrFile.exists()) {
-      jfrFile = File.createTempFile("otel-inferred-traces-", ".jfr");
+      jfrFile = File.createTempFile("otel-inferred-traces-", ".jfr", tempDir);
       jfrFile.deleteOnExit();
       canDeleteJfrFile = true;
     }
     if (activationEventsFile == null || !activationEventsFile.exists()) {
-      activationEventsFile = File.createTempFile("otel-inferred-activation-events-", ".bin");
+      activationEventsFile =
+          File.createTempFile("otel-inferred-activation-events-", ".bin", tempDir);
       activationEventsFile.deleteOnExit();
       canDeleteActivationEventsFile = true;
     }
@@ -352,6 +357,10 @@ public class SamplingProfiler implements Runnable {
   @Override
   @SuppressWarnings("FutureReturnValueIgnored")
   public void run() {
+    if (!config.isEnabled()) {
+      logger.fine("Profiling is disabled, not starting profiling session");
+      return;
+    }
 
     // lazily create temporary files
     try {